backend/requirements.txt
Flask
python-dotenv
numpy
pandas
flask-cors
scikit-learn
statsmodels
xgboost
tensorflow
yfinance
feedparser
pydantic
openai
httpx
lightgbm
